# resources/openapi/tabou2-api.json
{
	"openapi": "3.0.3",
	"info": {
		"description": "You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
		"version": "1.0.0",
		"title": "tabou2",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "apiteam@swagger.io"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"servers": [
		{
			"url": "/{basePath}",
			"description": "Tabou2 v2 API",
			"variables": {
				"basePath": {
					"default": "/v2"
				}
			}
		}
	],
...
	"paths": {
...
		"/types-evenements": {
			"get": {
				"summary": "Recherche les types d'évènements",
				"operationId": "searchTypesEvenements",
				"parameters": [
					{
						"in": "query",
						"name": "libelle",
						"description": "Recherche par libellé du type d'évènement",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "inactif",
						"description": "Recherche par type d'évènement inactif",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"name": "start",
						"in": "query",
						"required": false,
						"description": "Index de début - démarre à 0",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "resultsNumber",
						"in": "query",
						"required": false,
						"description": "Nombre de résultats à remonter",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "orderBy",
						"in": "query",
						"required": false,
						"description": "Nom du champ sur lequel trier les résultats",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "asc",
						"in": "query",
						"required": false,
						"description": "Indique si le tri doit être fait par ordre ascendant",
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/PageResult"
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Ajout d'un type d'évènements",
				"operationId": "createTypeEvenements",
				"requestBody": {
					"description": "Nouveau type d'évènement",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "tabou2-model.json#/components/schemas/TypeEvenement"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/TypeEvenement"
								}
							}
						}
					}
				}
			},
			"put": {
				"summary": "Modification d'un type d'évènements",
				"operationId": "updateTypeEvenements",
				"requestBody": {
					"description": "Type d'évènement à modifier",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "tabou2-model.json#/components/schemas/TypeEvenement"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/TypeEvenement"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/types-evenements/{typeEvenementId}": {
			"get": {
				"summary": "Récupération du détail d'un type d'évènement",
				"operationId": "getTypeEvenementById",
				"parameters": [
					{
						"in": "path",
						"name": "typeEvenementId",
						"description": "Identifiant du type d'évènement",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/TypeEvenement"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/types-evenements/{typeEvenementId}/inactivate": {
			"put": {
				"summary": "Rends un type d'évènement inactif",
				"operationId": "inactivateTypeEvenement",
				"parameters": [
					{
						"in": "path",
						"name": "typeEvenementId",
						"description": "Identifiant du type d'évènement",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/TypeEvenement"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
...
	}
}